pyproject.toml
[tool.poetry]
name = "bsapmd"
version = "0.1.0"
description = "BeSmartAndPro Monitoring Downloader"
authors = ["BeSmartAnd.Pro <hello@BeSmartAnd.Pro>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
requests = "^2.32.3"
PyYAML = "^6.0.2"

[tool.poetry.group.dev.dependencies]
poetry = "^1.8.3"
pyinstaller = "^6.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-pyinstaller-plugin.scripts]
bsapmd = { source = "bsapmd/main.py", type = "onefile", bundle = false, add_version=true }